Fix defect in sound-controller sound loading.
Amend 080b75e: `shouldPlayOnLoad` flag needs to be
reset not when starting play process, which is
before loading finishes, but right before playing
starts. Otherwise, when first calling to load the
sound, and then calling to play it, the sound can
play twice because when the first
player-item-replace finishes, the flag has already
been corrupted.